Update ST_D_L.py
Changed the D_L, which is the independent variable, to the shorthand term EspFact, for expansion factor. The expansion factor is the independent term used in the FLRW model of the Universe expansion. This will allow cosmologists and students quick understanding of the code when made public.  The code "connecting" the databases to the six models may have to be edited to reflect this change.  Also note have added another parentheses around the term (speed_of_light/(Hubble*EspFact*np.sqrt(np.fabs(1-Matter)))) also to enable easier student understanding of code. Have added some comments below, at lines 16 thru 30,  this long and complete function to aid understanding.